% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageReplicates.R
\name{averageReplicates}
\alias{averageReplicates}
\title{Average Assays Across Replicates}
\usage{
averageReplicates(list, assay.type = "matrix")
}
\arguments{
\item{list}{A list of \code{\link{SummarizedExperiment}} objects.}

\item{assay.type}{Name the assay to average between all samples.}
}
\value{
A \code{SummarizedExperiment} with the averaged assay.
}
\description{
This function takes a list of \code{SummarizedExperiment} objects.
It averages the requested assay between all input objects and returns
a new single \code{SummarizedExperiment} with the averaged assay.
}
\examples{
# Prepare example data ----
example(makeExperimentFromDeeptools)

# Make a list of samples
sampleNames <- head(LETTERS, 2)
binCenters <- seq_len(nWindows) - nWindows/2
seList <- lapply(sampleNames, function(x){makeExperimentFromDeeptools(tf, col.names=binCenters)})
names(seList) <- sampleNames
# Usage ----
averageReplicates(seList)
}
