% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binBreaks.R
\name{binBreaks}
\alias{binBreaks}
\alias{binCenters}
\title{Compute Boundaries and Center}
\usage{
binBreaks(from, to, width)

binCenters(from, to, width)
}
\arguments{
\item{from}{Start coordinate of the windows.}

\item{to}{End coordinate of the windows.}

\item{width}{Width of the windows.}
}
\value{
For \emph{n} bins:
\describe{
\item{\code{binBreaks}}{A numeric vector of length \emph{n+1} representing the breaks between bins.}
\item{\code{binCenters}}{A numeric vector of length \emph{n} representing the center of each bin.}
}
}
\description{
\code{binBreaks} computes the breaks of tiled windows over a range.
It is equal to the base function \code{seq()}.

\code{binCenters} computes the centre of tiled windows over a range.
}
\examples{
binBreaks(-500, 500, 10)
binCenters(-500, 500, 10)
}
