% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeperTools-package.R
\docType{package}
\name{deeperTools-pkg}
\alias{deeperTools}
\alias{deeperTools-package}
\alias{deeperTools-pkg-package}
\title{deeperTools: Downstream Analysis for deepTools}
\description{
\code{deeperTools} is a package that provides convenience functions to process files produced by deepTools \url{https://deeptools.readthedocs.io/en/develop}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinrue/deeperTools}
  \item Report bugs at \url{https://github.com/kevinrue/deeperTools/issues}
}

}
\author{
Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}
}
\keyword{internal}
