% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDeeptoolsSamples.R
\name{generateDeeptoolsExperiments}
\alias{generateDeeptoolsExperiments}
\title{Generate SummarizedExperiment Objects Representing deepTools Matrices}
\usage{
generateDeeptoolsExperiments(ranges, bins, names = c("A"))
}
\arguments{
\item{ranges}{Number of genomic ranges (i.e., rows).}

\item{bins}{Number of bins (i.e., columns).}

\item{names}{Character vector of sample names.}
}
\value{
A list of `SummarizedExperiment` objects.
}
\description{
Generate SummarizedExperiment Objects Representing deepTools Matrices
}
\examples{
# Usage ----

sample_names <- c("A", "B", "C")

se_list <- generateDeeptoolsExperiments(20, 10, sample_names)
se_list
}
