% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDeeptoolsExperiment.R
\name{importDeeptoolsExperiment}
\alias{importDeeptoolsExperiment}
\title{Import a deepTools Matrix File as a SummarizedExperiment Object}
\usage{
importDeeptoolsExperiment(file, col.names = NULL)
}
\arguments{
\item{file}{A gzip-compressed matrix file.}

\item{col.names}{A character vector of column names immediately set on the output object.}
}
\value{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object.
See \emph{Details}.
}
\description{
Import a deepTools Matrix File as a SummarizedExperiment Object
}
\details{
Matrix files produced by \href{https://deeptools.readthedocs.io/en/develop/}{deepTools} do not include column names that are useful to indicate the genomic position relative to a reference point, for instance.
While \code{colnames} could be manually set by users on the \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} returned, it can be convenient to provide a set of column names directly to the \code{\link[=importDeeptoolsExperiment]{importDeeptoolsExperiment()}} function, especially when importing multiple samples in an \code{\link[=lapply]{lapply()}} statement, for instance.

The returned object contains:
\itemize{
\item The deepTools matrix in the \code{assay} slot.
\item The genomic range information in the \code{rowRanges} slot.
}
}
\examples{
library(rtracklayer)
library(GenomicRanges)

# Prepare example data ----

se_list <- generateDeeptoolsExperiments(ranges=20, bins=10, names=c("A"))
se_list

# Write example file ----

se_A <- se_list[["A"]]
gr_A <- rowRanges(se_A)
df_A <- data.frame(
  seqnames(gr_A), start(gr_A), end(gr_A), names(gr_A), 0, strand(gr_A),
  assay(se_A))

# Write the sample data to file
tf <- tempfile(fileext=".matrix.gz")
conn <- gzfile(tf, "wt")
write.table(df_A, conn, row.names=FALSE, col.names=TRUE)
close(conn)

# Usage ----

importDeeptoolsExperiment(tf)
}
