% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binBreaks.R
\name{.check_to_from_width}
\alias{.check_to_from_width}
\title{Internal Validity Check of Function Arguments}
\usage{
.check_to_from_width(from, to, width)
}
\arguments{
\item{from}{Start coordinate of the windows.}

\item{to}{End coordinate of the windows.}

\item{width}{Width of the windows.}
}
\value{
\code{TRUE} if there are no issues. Otherwise, an error is thrown.
}
\description{
Internal Validity Check of Function Arguments
}
\keyword{internal}
