% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExperimentFromDeeptools.R
\name{makeExperimentFromDeeptools}
\alias{makeExperimentFromDeeptools}
\title{Make a SummarizedExperiment object from a deepTools matrix}
\usage{
makeExperimentFromDeeptools(file, col.names = NULL)
}
\arguments{
\item{file}{A gzip-compressed matrix file.}

\item{col.names}{}
}
\value{
A \code{\linkS4class{SummarizedExperiment}} object. See \emph{Details}.
}
\description{
Make a SummarizedExperiment object from a deepTools matrix
}
\details{
The returned object contains:
\itemize{
\item The deepTools matrix in the \code{assay} slot.
\item The genomic range information in the \code{rowRanges} slot.
}
}
\examples{
require(rtracklayer)
require(GenomicRanges)

# Prepare example data ----

# Make sample genomic ranges
nRanges <- 3
gr <- GRanges(
    seqnames = rep("chr1", nRanges),
    ranges = IRanges(seq(1, nRanges), 10+seq(1, nRanges)))
names(gr) <- letters[1:3]

# Make a sample matrix
nWindows <- 10
mat <- matrix(data = rbinom(nRanges*nWindows, 10, 0.1), nrow = nRanges)

# Combine the two
mcols(gr) <- mat

# Write the sample data to file
tf <- tempfile(fileext = ".matrix.gz")
conn <- gzfile(tf, "wt")
outData <- data.frame(seqnames(gr), start(gr), end(gr), names(gr), 0, strand(gr), mcols(gr))
write.table(outData, conn, row.names = FALSE, col.names = TRUE)
close(conn)

# Usage ----
binCenters <- seq_len(nWindows) - nWindows/2
se <- makeExperimentFromDeeptools(tf, col.names=binCenters)
}
