% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByGeneSet.R
\name{splitByGeneSet}
\alias{splitByGeneSet}
\title{Extracts Gene Subsets from a SummarizedExperiment object}
\usage{
splitByGeneSet(se, genesets)
}
\arguments{
\item{se}{A \code{\link{SummarizedExperiment}} object.}

\item{genesets}{A list of gene sets.}
}
\value{
A list of \code{\link{SummarizedExperiment}} objects.
}
\description{
Extracts Gene Subsets from a SummarizedExperiment object
}
\examples{
# Prepare example data ----

example(makeExperimentFromDeeptools)

# Usage ----

genesets <- list(set1=c("a"), set2=c("b", "c"))
sse <- splitByGeneSet(se, genesets)
}
