% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowCenters.R
\name{windowCenters}
\alias{windowCenters}
\alias{windowBreaks}
\title{Compute Window Centres}
\usage{
windowCenters(from, to, by)

windowBreaks(from, to, by)
}
\arguments{
\item{from}{Start coordinate of the windows.}

\item{to}{End coordinate of the windows.}

\item{by}{Width of the windows.}
}
\value{
A numeric vector.
}
\description{
Compute the centre of tiled windows over a range.

Compute the breaks of tiled windows over a range.
Identical to the base function \code{seq(from, to, by)}.
}
\examples{
windowCenters(-500, 500, 10)
windowBreaks(-500, 500, 10)
}
