% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowCenters.R
\name{windowCenters}
\alias{windowCenters}
\title{Compute the centre of windows}
\usage{
windowCenters(from, to, by)
}
\arguments{
\item{from}{Start coordinate of the windows.}

\item{to}{End coordinate of the windows.}

\item{by}{Width of the windows.}
}
\value{
A numeric vector.
}
\description{
Compute the centre of windows
}
\examples{
windowCenters(-500, 500, 10)
}
