% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageReplicates.R
\name{averageReplicates}
\alias{averageReplicates}
\title{Average Assays Across Replicates}
\usage{
averageReplicates(list, assay.type = "matrix")
}
\arguments{
\item{list}{A list of \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} objects.}

\item{assay.type}{Name the assay to average between all samples.}
}
\value{
A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object with the averaged assay.
}
\description{
This function takes a list of \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} objects.
It averages the requested assay between all input objects and returns a new single \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object with the averaged assay.
}
\examples{
# Prepare example data ----

sample_names <- c("A", "B", "C")

se_list <- generateDeeptoolsExperiments(20, 10, sample_names)
se_list

# Usage ----

averageReplicates(se_list)
}
