% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageReplicates.R
\name{averageReplicates}
\alias{averageReplicates}
\title{Average Assays Across Replicates}
\usage{
averageReplicates(list, assay.type = "matrix")
}
\arguments{
\item{list}{A list of \code{\link{SummarizedExperiment}} objects.}

\item{assay.type}{Name the assay to average between all samples.}
}
\value{
A \code{SummarizedExperiment} with the averaged assay.
}
\description{
This function takes a list of \code{SummarizedExperiment} objects.
It averages the requested assay between all input objects and returns
a new single \code{SummarizedExperiment} with the averaged assay.
}
\examples{
# Prepare example data ----

se_list <- generateDeeptoolsExperiments(20, 10, head(LETTERS, 3))

# Usage ----

averageReplicates(se_list)
}
