% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExperimentFromDeeptools.R
\name{makeExperimentFromDeeptools}
\alias{makeExperimentFromDeeptools}
\title{Make a SummarizedExperiment Object from a deepTools Matrix File}
\usage{
makeExperimentFromDeeptools(file, col.names = NULL)
}
\arguments{
\item{file}{A gzip-compressed matrix file.}

\item{col.names}{A character vector of column names immediately set on the output object.}
}
\value{
A \code{\linkS4class{SummarizedExperiment}} object. See \emph{Details}.
}
\description{
Make a SummarizedExperiment Object from a deepTools Matrix File
}
\details{
Matrix files produced by the \href{https://deeptools.readthedocs.io/en/develop/}{deepTools}
do not include column names that are useful to indicate the genomic position relative to a reference point, for instance.
While \code{colnames} could be manually set by users on the \code{SummarizedExperiment} returned,
it can be convenient to provide a set of column names directly to the \code{makeExperimentFromDeeptools} function,
especially when importing multiple samples in an \code{\link{lapply}} statement, for instance.

The returned object contains:
\itemize{
\item The deepTools matrix in the \code{assay} slot.
\item The genomic range information in the \code{rowRanges} slot.
}
}
\examples{
library(rtracklayer)
library(GenomicRanges)

# Prepare example data ----

n_bins <- 10

se_list <- generateDeeptoolsExperiments(ranges=20, bins=n_bins, names=c("A"))
se_A <- se_list[["A"]]

# Write example file ----

gr_A <- rowRanges(se_A)
df_A <- data.frame(
  seqnames(gr_A), start(gr_A), end(gr_A), names(gr_A), 0, strand(gr_A),
  assay(se_A, "matrix"))

# Write the sample data to file
tf <- tempfile(fileext=".matrix.gz")
conn <- gzfile(tf, "wt")
write.table(df_A, conn, row.names=FALSE, col.names=TRUE)
close(conn)

# Usage ----

bin_centers <- seq_len(n_bins) - n_bins/2
makeExperimentFromDeeptools(tf, col.names=bin_centers)
}
