% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMetageneDataFrame.R
\name{makeMetageneDataFrame}
\alias{makeMetageneDataFrame}
\title{Compute Metagenes for Lists of Samples and Gene Sets}
\usage{
makeMetageneDataFrame(list, assay.type = "matrix")
}
\arguments{
\item{list}{A named list of lists of \code{\link{SummarizedExperiment}} objects.
Each item in the top-most list is a named sample (e.g. "replicate1").
Each item within each sublist is a named gene set (e.g. "upregulated").
Each \code{SummarizedExperiment} stores matrix of histone modification enrichment,
with \code{colnames} set to the numerical value of the center of each window used to compute enrichment.}

\item{assay.type}{Name the assay to average between all samples.}
}
\value{
A \code{data.frame} with columns \code{c("sample", "geneset", "position")}
}
\description{
This function takes a list of data sets stratified by samples and gene sets.
It computes the metagene for each gene set in each sample, and returns
a \code{data.frame} suitable for \code{ggplot2}.
}
\examples{
# Set the plotting environment ----

library(ggplot2)
library(cowplot)
theme_set(theme_cowplot())

# Prepare example data ----

se_list <- generateDeeptoolsExperiments(20, 10, head(LETTERS, 3))

# Split each sample into gene subsets ----
range_sets <- list(set1=c("GR_1"), set2=c("GR_2", "GR_3"))
se_list_list <- lapply(se_list, splitByGeneSet, range_sets)

# Usage ----

x <- makeMetageneDataFrame(se_list_list)

ggplot(x, aes(position, metagene, color=geneset)) +
    geom_line(aes(group=interaction(geneset, sample))) +
    geom_point() +
    facet_wrap(~sample, ncol=1)
}
\seealso{
importDeeptoolsExperiment
}
