% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByGeneSet.R
\name{splitByGeneSet}
\alias{splitByGeneSet}
\title{Extracts Gene Subsets from a SummarizedExperiment object}
\usage{
splitByGeneSet(se, genesets)
}
\arguments{
\item{se}{A \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} object.}

\item{genesets}{A list of gene sets.}
}
\value{
A list of \link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment} objects.
}
\description{
Extracts Gene Subsets from a SummarizedExperiment object
}
\examples{
# Prepare example data ----

sample_names <- c("A", "B")

se_list <- generateDeeptoolsExperiments(20, 10, sample_names)
se_list

# Usage ----

range_sets <- list(set1=c("GR_1"), set2=c("GR_2", "GR_3"))
splitByGeneSet(se_list[[1]], range_sets)
}
